backend/requirements.txt
flask==2.3.3
flask-cors==3.0.10
openai==0.27.0
numpy==2.2.3
python-dotenv==0.19.0
werkzeug==2.3.7
